requirements.txt
apriori-python         == 1.0.4
efficient-apriori      == 2.0.5
fpgrowth-py            == 1.0.0
pyarmviz               == 0.1.4
pandas~=2.2.3
ucimlrepo              == 0.0.7
scikit-learn           == 1.5.2
matplotlib~=3.9.2
seaborn~=0.13.2
trimap                 == 1.1.4
pacmap                 == 0.7.3
umap-learn             == 0.5.6

efficient_apriori~=2.0.5
pillow~=10.4.0
Jinja2~=3.1.4
numpy~=2.1.3
pip~=23.2.1
scipy~=1.14.1